Began working on base layout, bootstrap
Made file structure and started to work on going through
and adding the needed structure to make the website
uniform across all pages. Note that there are only currently
placeholder values for most all of the elements and nothing
is really solid. From here on the image folder will be added
to the gitignore list.