JS for gallery thumbnails
Added in JS for gallery thumbnails. Works only with odd number of thumbnails,
completely breaks with even numbers. Odd numbers is better because it will allow
the middle one to be the one selected when possible, making it an easier
experience overall. Also works with any number of thumbnails and any odd number
of of max thumbnails shown.