Finished all classes, refactored
Added a lot of changes and bug fixes. Added the ability to add
classes to thumbnails, and to remove classes. set gallery and
event listeners now work as expected. Can now set custom ids for
the main image and thumbnailHolder. Fixed a small screen size
bug. Added a couple helper functions inside of methods to help
code be a little bit more readable. Fixed a bug where the src
of stored imgs was give the absolute path and comparing that
instead of comparing the relative path which is stored.

TODO update README and asked for code to be looked over